Extract `WorkerSharedUserHomeScopeServices`
With the services shared between `GradleUserHomeScopeServices` and
`WorkerDaemonServices`.
